Separate GUI: Moved some Tkinter code into gui.py
